move enum to card types
