+++ crates/opsml_sql/src/sqlite/helper.rs
@@ -335,6 +335,7 @@ impl SqliteQueryHelper {
         AND (?2 IS NULL OR space = ?2)
         AND (?3 IS NULL OR name = ?3)
         AND (?4 IS NULL OR created_at <= DATETIME(?4))
+        AND (?5 IS NULL OR service_type = ?5)
         "
         );
