+++ crates/opsml_sql/src/sqlite/sql/card/mod.rs
@@ -177,7 +177,8 @@ impl CardLogicTrait for CardLogicSqliteClient {
                     .bind(query_args.space.as_ref())
                     .bind(query_args.name.as_ref())
                     .bind(query_args.max_date.as_ref())
-                    .bind(query_args.limit.unwrap_or(50))
+                    .bind(query_args.service_type.as_ref())
+                    .bind(query_args.limit.unwrap_or(1000))
                     .fetch_all(&self.pool)
                     .await?;
 
@@ -358,6 +359,7 @@ impl CardLogicTrait for CardLogicSqliteClient {
                         .bind(&record.cards)
                         .bind(&record.username)
                         .bind(&record.opsml_version)
+                        .bind(&record.service_type)
                         .execute(&self.pool)
                         .await?;
                     Ok(())
